Semi 286 refactor new creditcredential

See merge request atixlabs-oss/semillas-middleware!86

+++ src/main/java/com/atixlabs/semillasmiddleware/app/dto/CredentialPage.java
@@ -19,9 +19,9 @@ import java.time.LocalDateTime;
 public class CredentialPage {
 
    Page<CredentialDto> credentialsDto;
-   Long amountOfElements;
+   Integer amountOfElements;
 
-    public CredentialPage(Page<CredentialDto> pageDto, Long totalAmountOfItems) {
+    public CredentialPage(Page<CredentialDto> pageDto, Integer totalAmountOfItems) {
         this.credentialsDto = pageDto;
         this.amountOfElements = totalAmountOfItems;
     }
